scripts/requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
emoji>=2.0.0
numpy>=1.24.0
